rate list of people, add fonts, move license around, add graphview library

+++ app/src/main/java/com/osacky/hipsterviz/LoadingInterface.java
@@ -3,7 +3,7 @@ package com.osacky.hipsterviz;
 public interface LoadingInterface {
 
     public abstract void onLoadingStarted();
-    public abstract void onLoadingProgressUpdate(int progress);
+    public abstract void onLoadingProgressUpdate(float progress);
     public abstract void onLoadingFinished();
 
 }
